Moves AI model to directory
Moves AI model to directory.
Minor code refactor.
