--- tapas_model/model.py
@@ -1,6 +1,0 @@
-
-# # Load model directly
-# from transformers import AutoTokenizer, AutoModelForTableQuestionAnswering
-
-# tokenizer = AutoTokenizer.from_pretrained("google/tapas-base-finetuned-wtq")
-# model = AutoModelForTableQuestionAnswering.from_pretrained("google/tapas-base-finetuned-wtq")
